Increased size of disk buffer to reduce SD card IO
Changed a number of hard coded numbers relating to the SD disks with
defines to that the number of bytes in a segment could be changed to
reduce the number of SD card reads needed to service a disk operation.

Currently the buffer is set to 128, it could be set higher by changing
SEGMENT_SIZE, max value is 512 for a 1-2-1 mapping between segment and
sector.