Fix issues with new SD library
The new SD library always creates file opened for write meaning that a
blank binary file was created if none existed when looking for new disk
sets. This meant that the TXT file was ignored. Updated the wrapper to
check to see if file exists before opening.

Also added saving TXT file contents as a DEF file to speed loading next
time.